final/requirements.txt
navec==0.10.0
pytest==8.4.2
PyYAML==6.0.2
setuptools==78.1.1
torch==2.8.0
